+++ account-api/src/main/java/org/healthplus/account/application/AccountService.java
@@ -4,11 +4,11 @@ import lombok.extern.slf4j.Slf4j;
 import org.healthplus.account.application.command.SigninCommand;
 import org.healthplus.account.application.command.SignupCommand;
 import org.healthplus.account.application.result.AccountResult;
-import org.healthplus.account.domain.entity.User;
+import org.healthplus.account.domain.User;
 import org.healthplus.account.domain.exception.EmailInfoMisMatchException;
 import org.healthplus.account.domain.exception.PasswordMisMatchException;
-import org.healthplus.account.domain.repository.EncryptMapper;
-import org.healthplus.account.domain.repository.UserRepository;
+import org.healthplus.account.domain.EncryptMapper;
+import org.healthplus.account.domain.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
